Move loggers and summary writers to a separate package
