+++ src/trainer/log_utils/__init__.py
@@ -1,0 +1,2 @@
+from .logger import get_logger
+from .writer import get_summary_writer
