+++ src/trainer/trainer.py
@@ -6,78 +6,19 @@ from abc import abstractmethod
 
 # Helper libraries
 from copy import deepcopy
-from logging import Logger, getLogger, FileHandler, StreamHandler, Formatter
-import logging
-import shutil
 
 # Machine learning
 import torch
 import torch.nn as nn
 from torch.utils.data import DataLoader
-from torch.utils.tensorboard import SummaryWriter
 from tqdm import tqdm
 import numpy as np
 import pandas as pd
 
-# --- Logging ---
+# Logging
+from logging import Logger
+from .log_utils import get_logger, get_summary_writer
 
-def get_logger(name: str = 'ml-sandbox') -> Logger:
-    """Create a logging object in the debugging format.
-
-    Args:
-        name (str, optional): Name of the object. Defaults to 'ml-sandbox'.
-
-    Returns:
-        Logger: An instance of a logger.
-    
-    Example:
-        >>> # Use the logger
-        >>> logger.debug('This is a debug message')
-        >>> logger.info('This is an info message')
-        >>> logger.error('This is an error message')
-    """
-    # Create logger
-    logger = getLogger(name)
-    logger.setLevel(logging.DEBUG)
-
-    # Create console handler
-    console_handler = StreamHandler()
-    console_handler.setLevel(logging.INFO)
-
-    # # Create file handler
-    # file_handler = FileHandler('app.log')
-    # file_handler.setLevel(logging.DEBUG)
-
-    # Create formatter
-    formatter = Formatter(
-        '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
-    )
-
-    # Add formatter to handlers
-    console_handler.setFormatter(formatter)
-    # file_handler.setFormatter(formatter)
-
-    # Add handlers to logger
-    logger.addHandler(console_handler)
-    # logger.addHandler(file_handler)
-
-    return logger
-
-# --- TensorBoard Summary Writer ---
-
-def get_summary_writer(run_dir: str = "logs", delete_logs=True) -> SummaryWriter:
-    """Get a base TensorBoard summary writer.
-
-    Args:
-        run_dir (str): The directory where the TensorBoard logs will be saved.
-
-    Returns:
-        SummaryWriter: A TensorBoard summary writer.
-    """
-    if delete_logs:
-        shutil.rmtree(run_dir, ignore_errors=True)
-    
-    return SummaryWriter(log_dir=run_dir)
 
 # --- Utility functions ---
